VERSION 1.3
An open-source tool for flashing LPC microcontrollers. Currently
tested on Ubuntu 20.04 (64-bit) with Qt 5.15 on LPC1343 and LPC1347.
Should also work on LPC1315, LPC1316, LPC1317, LPC1345, LPC1346
because they use the same bootloader, but it is not tested!
If you tried any of these, contact me on lbogdanov@tu-sofia.bg
for feedback.

This program is free and can be used for open and closed source
projects. I'd be happy if you tell me for what you are using it,
so contact me on lbogdanov@tu-sofia.bg